fix(localization): replaced Int type on Long for number parameter in getStringWithOneNumber() method in LocalizationFormatter

+++ app/src/main/kotlin/org/coremapx/app/localization/objects/LocalizationFormatter.kt
@@ -28,7 +28,7 @@ object LocalizationFormatter {
             .replace(NEXT_LINE, "\n")
     }
 
-    fun getStringWithOneNumber(startString: String, number: Int): String {
+    fun getStringWithOneNumber(startString: String, number: Long): String {
         return startString
             .replace(NUMBER, number.toString())
     }
